Fix compiler warning (possible loss of data)
Fix 306 warnings from MS C:

tesseract\ccutil\unicharset.h(242): warning C4267:
 'argument': conversion from 'size_t' to 'int', possible loss of data

The change also avoids some type conversions.

+++ ccutil/unicharset.h
@@ -241,7 +241,7 @@ class UNICHARSET {
   static std::string CleanupString(const char* utf8_str) {
     return CleanupString(utf8_str, strlen(utf8_str));
   }
-  static std::string CleanupString(const char* utf8_str, int length);
+  static std::string CleanupString(const char* utf8_str, size_t length);
 
   // Return a STRING containing debug information on the unichar, including
   // the id_to_unichar, its hex unicodes and the properties.
